#71 Split evaluator
