+++ src/main/scala/org/sruby/evaluator/EvaluatorLike.scala
@@ -1,0 +1,8 @@
+package org.sruby.evaluator
+
+import org.sruby.parser.AST
+
+trait EvaluatorLike {
+  def eval: PartialFunction[AST, Evaluator.Evaluation]
+}
+
