+++ src/main/scala/org/sruby/evaluator/RubyMethod.scala
@@ -5,16 +5,13 @@ import org.sruby.parser.AST
 import org.sruby.core.RubyObject
 
 class RubyMethod(val params: Seq[Symbol], val ts: List[AST]) {
-  def invoke(
-    args: State[EvalState, Seq[RubyObject]],
-    evals: List[AST] => State[EvalState, RubyObject]
-  ): State[EvalState, RubyObject] = args.flatMap { evaldArgs =>
+  def invoke(args: State[EvalState, Seq[RubyObject]]): State[EvalState, RubyObject] = args.flatMap { evaldArgs =>
     evaldArgs.length == params.length match {
       case true =>
         val argMap = (params zip evaldArgs).toMap
         val tsWithBoundvars = ts.map(_.withBoundVars(argMap))
 
-        evals(tsWithBoundvars)
+        Evaluator.evalList(tsWithBoundvars)
       case false =>
         val errorMessage =
           s"wrong number of arguments (given ${evaldArgs.length}, expected ${params.length})"
